+++ tests/Agent/IntegrationTests/IntegrationTests/BasicInstrumentation/MvcAsyncTests.cs
@@ -11,7 +11,7 @@ using NewRelic.Testing.Assertions;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.IntegrationTests
+namespace NewRelic.Agent.IntegrationTests.BasicInstrumentation
 {
     [NetFrameworkTest]
     public class MvcAsyncTests : IClassFixture<MvcAsyncFixture>
